Added entityIcons to client combatUI
Client combatUI now shows entities' icons. If one is not found it falls back on default images.
Dynamic ability generation is still buggy.
Seems that if staff is equipped before another item like armor the abilities show correctly. But if done the other way the abilities granted by staff are not generated, despite showing in playerData.